Some changes.
The engine can be compiled successfully now. Some changes to the network
are still left to be done. For other parts, maybe there are also something
I havn't realized left but I think most of them are done though. In the
coming days, I will focus on modifying the tests and samples.